Crate diem_management

Modules

config

constants

error

secure_backend

storage

transaction

validator_config

waypoint

Macros

execute_command

secure_backend

Functions

read_key_from_file
Reads a given ed25519 public key from file. Attempts to read the key using bcs encoding first. If this fails, attempts reading the key using hex.